requirements.txt
psycopg2==2.8.6
SQLAlchemy==1.4.17
python-telegram-bot==13.4.1

